bot/pyproject.toml
[tool.poetry]
name = "chatbot_sa"
version = "0.1.0"
description = "Telegram bot with FastAPI"
authors = ["Noam Salomonski"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
python-telegram-bot = {extras = ["webhooks"], version = "^20.7"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
